face_battle/requirements.txt
requests
pydantic
loguru
slack-bolt
openai
Pillow
litellm
jinja2